Feature/create user crud endpoints

+++ backend/src/main/java/com/example/backend/token/TokenService.java
@@ -5,6 +5,6 @@ import com.example.backend.user.User;
 public interface TokenService {
     public abstract Boolean isTokenValid(String token);
     public abstract void saveTokenByUser(String token, User user);
-    public abstract void revokeAllTokensByUserId(int id);
+    public abstract void revokeAllTokensByUserId(Long id);
     public abstract void revokeTokenByJwt(String token);
 }
